#10 Functions optimization, reuse of new entities
Mapping.py functions have been rewritten to ensure efficient handling of newly created subrecords. Recursion has been improved by a new way of expressing subrecord links in the front-end environment. Main.js was modified for this purpose and to correct the function dedicated to editing subrecords.
It is now possible to reuse a subrecord entity within the same record form.